django/requirements.txt
django==5.1.2
flake8==6.1.0
pydantic_settings==2.5.2
django-split-settings==1.3.2
psycopg[binary]==3.1.18
gunicorn==21.2.0
uvicorn==0.31.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
loguru==0.7.2
django-bootstrap5==24.3
django-spurl==0.6.8
crispy-bootstrap5==2024.2
django-crispy-forms==2.3